backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.0.1
openai==1.51.2
pydantic==2.9.2r